mcp-server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mcp==1.13.0
openai==1.99.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
sseclient==0.0.27
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
